+++ hal/src/photon/socket_hal.cpp
@@ -31,6 +31,7 @@
 #include <algorithm>
 #include <vector>
 #include "lwip/api.h"
+#include "network_interface.h"
 
 /**
  * Socket handles
@@ -796,7 +797,7 @@ sock_result_t socket_create_tcp_server(uint16_t port, network_interface_t nif)
     tcp_server_t* server = new tcp_server_t();
     wiced_result_t result = WICED_OUT_OF_HEAP_SPACE;
     if (handle && server) {
-        result = wiced_tcp_server_start(server, WICED_STA_INTERFACE,
+        result = wiced_tcp_server_start(server, wiced_wlan_interface(nif),
             port, WICED_MAXIMUM_NUMBER_OF_SERVER_SOCKETS, server_connected, server_received, server_disconnected, NULL);
     }
     if (result!=WICED_SUCCESS) {
@@ -878,10 +879,10 @@ sock_handle_t socket_create(uint8_t family, uint8_t type, uint8_t protocol, uint
         wiced_result_t wiced_result;
         socket->set_type((protocol==IPPROTO_UDP ? socket_t::UDP : socket_t::TCP));
         if (protocol==IPPROTO_TCP) {
-            wiced_result = wiced_tcp_create_socket(tcp(socket), WICED_STA_INTERFACE);
+            wiced_result = wiced_tcp_create_socket(tcp(socket), wiced_wlan_interface(nif));
         }
         else {
-            wiced_result = wiced_udp_create_socket(udp(socket), port, WICED_STA_INTERFACE);
+            wiced_result = wiced_udp_create_socket(udp(socket), port, wiced_wlan_interface(nif));
         }
         if (wiced_result!=WICED_SUCCESS) {
             socket->set_type(socket_t::NONE);
@@ -983,12 +984,12 @@ sock_result_t socket_join_multicast(const HAL_IPAddress *address, network_interf
 {
     wiced_ip_address_t multicast_address;
     SET_IPV4_ADDRESS(multicast_address, address->ipv4);
-    return as_sock_result(wiced_multicast_join(wiced_interface_t(nif), &multicast_address));
+    return as_sock_result(wiced_multicast_join(wiced_wlan_interface(nif), &multicast_address));
 }
 
 sock_result_t socket_leave_multicast(const HAL_IPAddress *address, network_interface_t nif, void * /*reserved*/)
 {
     wiced_ip_address_t multicast_address;
     SET_IPV4_ADDRESS(multicast_address, address->ipv4);
-    return as_sock_result(wiced_multicast_leave(wiced_interface_t(nif), &multicast_address));
+    return as_sock_result(wiced_multicast_leave(wiced_wlan_interface(nif), &multicast_address));
 }